backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
pydantic==2.8.2
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
jinja2==3.1.4
